apps/hlab/pyproject.toml
[project]
name = "hlab"
version = "0.1.0"
description = "Homelab monitoring and management CLI"
requires-python = ">=3.11"
dependencies = [
    "cells",
    "data",
    "vertex",
    "dsl",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[tool.uv.sources]
cells = { workspace = true }
data = { workspace = true }
vertex = { workspace = true }
dsl = { workspace = true }

[project.scripts]
hlab = "hlab.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hlab"]
